refactor: 공통 - controllers, services 일관화
