+++ controllers/threadController.js
@@ -1,7 +1,8 @@
-const { createThreadService } = require('../services/threadService');
-const { throwError } = require('../utils/throwError');
+const { threadServices } = require('../services');
+const { createThreadService } = threadServices;
+const { throwError } = require('../utils');
 
-exports.createThreadController = async (req, res, next) => {
+const createThreadController = async (req, res, next) => {
   try {
     const { id } = req.user;
     const { content } = req.body;
@@ -12,4 +13,8 @@ exports.createThreadController = async (req, res, next) => {
     console.error(err);
     next(err);
   }
+};
+
+module.exports = {
+  createThreadController,
 };
