+++ routes/threadRouter.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const { verifyToken } = require('../middlewares');
-const { createThreadController } = require('../controllers/threadController');
+const { threadControllers } = require('../controllers');
+const { createThreadController } = threadControllers;
 const router = express.Router();
 
 router.post('/', verifyToken, createThreadController);
