+++ routes/userRouter.js
@@ -1,12 +1,13 @@
 const express = require('express');
+const { userControllers } = require('../controllers');
 const {
+  signUpController,
+  loginController,
   getVerificationCodeController,
   setNewPasswordController,
-} = require('../controllers/resetPasswordController');
+} = userControllers;
 const { verifyToken } = require('../middlewares');
-const { signUpController } = require('../controllers/signUpController');
 const router = express.Router();
-const { loginController } = require('../controllers/loginController');
 
 router.post('/login', loginController);
 
@@ -14,4 +15,4 @@ router.post('/reset-password', getVerificationCodeController);
 router.post('/new-password', verifyToken, setNewPasswordController);
 router.post('/signup', signUpController);
 
-module.exports = router
+module.exports = router;
