+++ services/index.js
@@ -1,0 +1,19 @@
+const {
+  createUser,
+  getVerificationCodeService,
+  setNewPasswordService,
+  isEmailValid,
+} = require('./userService');
+const { createThreadService } = require('./threadService');
+
+module.exports = {
+  userServices: {
+    createUser,
+    getVerificationCodeService,
+    setNewPasswordService,
+    isEmailValid,
+  },
+  threadServices: {
+    createThreadService,
+  },
+};
