--- utils/tokenGeneration.js
@@ -1,6 +1,0 @@
-const jwt = require('jsonwebtoken');
-
-exports.tokenGeneration = (userId) => {
-  const token = jwt.sign({ id: userId }, process.env.JWT_SECRET);
-  return token;
-};
